# Conflicts:
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/queryexecutors/QueryTranslator.java
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/queryexecutors/creodias2/QueryExecutorCreoDias2.java
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/utils/WasdiFileUtils.java

+++ wasdiwebserver/src/main/java/it/fadeout/rest/resources/ProjectResource.java
@@ -12,7 +12,6 @@ import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Response;
@@ -21,13 +20,14 @@ import javax.ws.rs.core.Response.Status;
 import it.fadeout.Wasdi;
 import wasdi.shared.business.Project;
 import wasdi.shared.business.Subscription;
-import wasdi.shared.business.User;
+import wasdi.shared.business.users.User;
 import wasdi.shared.data.ProjectRepository;
 import wasdi.shared.data.SubscriptionRepository;
 import wasdi.shared.data.UserRepository;
 import wasdi.shared.utils.PermissionsUtils;
 import wasdi.shared.utils.Utils;
 import wasdi.shared.utils.log.WasdiLog;
+import wasdi.shared.viewmodels.ClientMessageCodes;
 import wasdi.shared.viewmodels.ErrorResponse;
 import wasdi.shared.viewmodels.SuccessResponse;
 import wasdi.shared.viewmodels.organizations.ProjectEditorViewModel;
@@ -35,11 +35,19 @@ import wasdi.shared.viewmodels.organizations.ProjectListViewModel;
 import wasdi.shared.viewmodels.organizations.ProjectViewModel;
 import wasdi.shared.viewmodels.organizations.SubscriptionListViewModel;
 
+
+/**
+ * Projects Resource.
+ * 	Hosts the api to let the user create new projects and select the active one.
+ * 		.create edit and delete projects
+ * 		.set active projects
+ * 		.read projects
+ * 
+ * @author p.campanella
+ *
+ */
 @Path("/projects")
 public class ProjectResource {
-
-	private static final String MSG_ERROR_INVALID_SESSION = "MSG_ERROR_INVALID_SESSION";
-
 	/**
 	 * Get the list of projects associated to a user.
 	 * @param sSessionId User Session Id
@@ -60,8 +68,8 @@ public class ProjectResource {
 
 		// Domain Check
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.getListByUser: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.getListByUser: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
 
 		try {
@@ -135,19 +143,28 @@ public class ProjectResource {
 
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
-		List<ProjectListViewModel> aoProjectList = new ArrayList<>();
-
 		// Domain Check
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.getListBySubscription: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.getListBySubscription: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
+		}
+		
+		SubscriptionRepository oSubscriptionRepository = new SubscriptionRepository();
+		Subscription oSubscription = oSubscriptionRepository.getSubscriptionById(sSubscriptionId);
+		
+		if (oSubscription == null) {
+			WasdiLog.warnLog("ProjectResource.getListBySubscription: invalid subscription");
+			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();						
+		}		
+		
+		if (!PermissionsUtils.canUserAccessSubscription(oUser.getUserId(), sSubscriptionId)) {
+			WasdiLog.warnLog("ProjectResource.getListBySubscription: user cannot access the subscription");
+			return Response.status(Status.FORBIDDEN).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();			
 		}
 
 		try {
-
-			SubscriptionRepository oSubscriptionRepository = new SubscriptionRepository();
-
-			Subscription oSubscription = oSubscriptionRepository.getSubscriptionById(sSubscriptionId);
+			
+			List<ProjectListViewModel> aoProjectList = new ArrayList<>();
 
 			String sSubscriptionName = null;
 
@@ -193,13 +210,13 @@ public class ProjectResource {
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.getProjectViewModel: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.getProjectViewModel: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
 
 		// Domain Check
 		if (Utils.isNullOrEmpty(sProjectId)) {
-			WasdiLog.debugLog("ProjectResource.getProjectViewModel: invalid project id");
+			WasdiLog.warnLog("ProjectResource.getProjectViewModel: invalid project id");
 			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("Invalid projectId.")).build();
 		}
 
@@ -217,7 +234,7 @@ public class ProjectResource {
 
 			if (!PermissionsUtils.canUserAccessSubscription(oUser.getUserId(), oProject.getSubscriptionId())) {
 				WasdiLog.debugLog("ProjectResource.getProjectViewModel: user cannot access project info, aborting");
-				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The user cannot access the project info.")).build();
+				return Response.status(Status.FORBIDDEN).entity(new ErrorResponse("The user cannot access the project info.")).build();
 			}
 
 			oVM = convert(oProject, oUser.getActiveProjectId());
@@ -239,34 +256,50 @@ public class ProjectResource {
 	@Path("/add")
 	@Produces({ "application/xml", "application/json", "text/xml" })
 	public Response createProject(@HeaderParam("x-session-token") String sSessionId, ProjectEditorViewModel oProjectEditorViewModel) {
+		
 		WasdiLog.debugLog("ProjectResource.createProject( Project: " + oProjectEditorViewModel.toString() + ")");
 
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.createProject: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.createProject: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
+	
+		try {
+			
+			if (!PermissionsUtils.canUserWriteSubscription(oUser.getUserId(), oProjectEditorViewModel.getSubscriptionId())) {
+				WasdiLog.warnLog("ProjectResource.createProject: invalid session");
+				return Response.status(Status.FORBIDDEN).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();				
+			}
+			
+			ProjectRepository oProjectRepository = new ProjectRepository();
 
-		ProjectRepository oProjectRepository = new ProjectRepository();
-
-		String sName = oProjectEditorViewModel.getName();
-		
-		while (oProjectRepository.getByName(sName) != null) {
-			sName = Utils.cloneName(sName);
-			WasdiLog.debugLog("ProjectResource.createProject: a Project with the same name already exists. Changing the name to " + sName);
-		}
-
-		Project oProject = convert(oProjectEditorViewModel);
-		oProject.setProjectId(Utils.getRandomName());
-
-		if (oProjectRepository.insertProject(oProject)) {
-			if (oProjectEditorViewModel.isActiveProject()) {
-				this.changeActiveProject(sSessionId, oProject.getProjectId());
+			String sName = oProjectEditorViewModel.getName();
+			
+			while (oProjectRepository.getByName(sName) != null) {
+				sName = Utils.cloneName(sName);
+				WasdiLog.debugLog("ProjectResource.createProject: a Project with the same name already exists. Changing the name to " + sName);
 			}
 
-			return Response.ok(new SuccessResponse(oProject.getProjectId())).build();
-		} else {WasdiLog.debugLog("ProjectResource.createProject( " + oProjectEditorViewModel.getName() + " ): insertion failed");
+			Project oProject = convert(oProjectEditorViewModel);
+			oProject.setProjectId(Utils.getRandomName());
+
+			if (oProjectRepository.insertProject(oProject)) {
+				if (oProjectEditorViewModel.isActiveProject()) {
+					this.changeActiveProject(sSessionId, oProject.getProjectId());
+				}
+
+				return Response.ok(new SuccessResponse(oProject.getProjectId())).build();
+			} 
+			else 
+			{
+				WasdiLog.debugLog("ProjectResource.createProject( " + oProjectEditorViewModel.getName() + " ): insertion failed");
+				return Response.serverError().build();
+			}			
+		}
+		catch (Exception oEx) {
+			WasdiLog.errorLog("ProjectResource.createProject: error ", oEx);
 			return Response.serverError().build();
 		}
 	}
@@ -286,51 +319,61 @@ public class ProjectResource {
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.updateProject: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.updateProject: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
+		
+		try {
+			ProjectRepository oProjectRepository = new ProjectRepository();
 
-		ProjectRepository oProjectRepository = new ProjectRepository();
+			Project oExistingProject = oProjectRepository.getProjectById(oProjectEditorViewModel.getProjectId());
 
-		Project oExistingProject = oProjectRepository.getProjectById(oProjectEditorViewModel.getProjectId());
-
-		if (oExistingProject == null) {
-			WasdiLog.debugLog("ProjectResource.updateProject: project does not exist");
-			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the Id exists.")).build();
-		}
-
-		Project oExistingProjectWithTheSameName = oProjectRepository.getByName(oProjectEditorViewModel.getName());
-
-		if (oExistingProjectWithTheSameName != null
-				&& !oExistingProjectWithTheSameName.getProjectId().equalsIgnoreCase(oExistingProject.getProjectId())) {
-			WasdiLog.debugLog("ProjectResource.updateProject: a different project with the same name already exists");
-			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("An project with the same name already exists.")).build();
-		}
-
-
-		Project oProject = convert(oProjectEditorViewModel);
-
-		if (oProjectRepository.updateProject(oProject)) {
-
-			if (oProjectEditorViewModel.isActiveProject()) {
-				this.changeActiveProject(sSessionId, oProject.getProjectId());
-			} else if (oProject.getProjectId().equals(oUser.getActiveProjectId())) {
-				UserRepository oUserRepository = new UserRepository();
-
-				oUser.setActiveProjectId(null);
-				oUser.setActiveSubscriptionId(null);
-
-				if (!oUserRepository.updateUser(oUser)) {
-					WasdiLog.debugLog("ProjectResource.updateProject( " + "changing the active project of the user to null failed");
-					return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The removing of the active project failed.")).build();
-				}
+			if (oExistingProject == null) {
+				WasdiLog.warnLog("ProjectResource.updateProject: project does not exist");
+				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the Id exists.")).build();
+			}
+			
+			if (!PermissionsUtils.canUserWriteSubscription(oUser.getUserId(), oExistingProject.getSubscriptionId())) {
+				WasdiLog.warnLog("ProjectResource.updateProject: user cannot write subscription ");
+				return Response.status(Status.FORBIDDEN).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();			
 			}
 
-			return Response.ok(new SuccessResponse(oProject.getProjectId())).build();
-		} else {
-			WasdiLog.debugLog("ProjectResource.updateProject( " + oProjectEditorViewModel.getName() + " ): update failed");
-			return Response.serverError().build();
+			Project oExistingProjectWithTheSameName = oProjectRepository.getByName(oProjectEditorViewModel.getName());
+
+			if (oExistingProjectWithTheSameName != null
+					&& !oExistingProjectWithTheSameName.getProjectId().equalsIgnoreCase(oExistingProject.getProjectId())) {
+				WasdiLog.warnLog("ProjectResource.updateProject: a different project with the same name already exists");
+				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("An project with the same name already exists.")).build();
+			}
+
+			Project oProject = convert(oProjectEditorViewModel);
+
+			if (oProjectRepository.updateProject(oProject)) {
+
+				if (oProjectEditorViewModel.isActiveProject()) {
+					this.changeActiveProject(sSessionId, oProject.getProjectId());
+				} else if (oProject.getProjectId().equals(oUser.getActiveProjectId())) {
+					UserRepository oUserRepository = new UserRepository();
+
+					oUser.setActiveProjectId(null);
+					oUser.setActiveSubscriptionId(null);
+
+					if (!oUserRepository.updateUser(oUser)) {
+						WasdiLog.warnLog("ProjectResource.updateProject( " + "changing the active project of the user to null failed");
+						return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The removing of the active project failed.")).build();
+					}
+				}
+
+				return Response.ok(new SuccessResponse(oProject.getProjectId())).build();
+			} else {
+				WasdiLog.debugLog("ProjectResource.updateProject( " + oProjectEditorViewModel.getName() + " ): update failed");
+				return Response.serverError().build();
+			}			
 		}
+		catch (Exception oEx) {
+			WasdiLog.errorLog("ProjectResource.updateProject: error ", oEx);
+			return Response.serverError().build();
+		}		
 	}
 
 	/**
@@ -348,37 +391,50 @@ public class ProjectResource {
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.changeActiveProject: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.changeActiveProject: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
+		
+		try {
+			String sSubscriptionId = null;
 
-		String sSubscriptionId = null;
+			if (sProjectId != null) {
+				ProjectRepository oProjectRepository = new ProjectRepository();
 
-		if (sProjectId != null) {
-			ProjectRepository oProjectRepository = new ProjectRepository();
+				Project oProject = oProjectRepository.getProjectById(sProjectId);
 
-			Project oProject = oProjectRepository.getProjectById(sProjectId);
+				if (oProject == null) {
+					WasdiLog.warnLog("ProjectResource.changeActiveProject: project does not exist");
+					return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the Id " + sProjectId + " exists.")).build();
+				}
 
-			if (oProject == null) {
-				WasdiLog.debugLog("ProjectResource.changeActiveProject: project does not exist");
-				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the Id " + sProjectId + " exists.")).build();
+				sSubscriptionId = oProject.getSubscriptionId();
+
+				if (!PermissionsUtils.canUserAccessSubscription(oUser.getUserId(), sSubscriptionId)) {
+					WasdiLog.warnLog("ProjectResource.changeActiveProject: user cannot access the subscription");
+					return Response.status(Status.FORBIDDEN).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();				
+				}
+				
 			}
 
-			sSubscriptionId = oProject.getSubscriptionId();
+			UserRepository oUserRepository = new UserRepository();
+
+			oUser.setActiveSubscriptionId(sSubscriptionId);
+			oUser.setActiveProjectId(sProjectId);
+
+			if (oUserRepository.updateUser(oUser)) {
+				WasdiLog.warnLog("ProjectResource.changeActiveProject: active project changed to " + sProjectId + " for user " + oUser.getUserId());
+				return Response.ok(new SuccessResponse(sProjectId)).build();
+			} 
+			else {
+				WasdiLog.debugLog("ProjectResource.changeActiveProject( " + "changing the active project of the user to " + sProjectId + " failed");
+				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The changing of the active project failed.")).build();
+			}			
 		}
-
-		UserRepository oUserRepository = new UserRepository();
-
-		oUser.setActiveSubscriptionId(sSubscriptionId);
-		oUser.setActiveProjectId(sProjectId);
-
-		if (oUserRepository.updateUser(oUser)) {
-			return Response.ok(new SuccessResponse(sProjectId)).build();
-		} 
-		else {
-			WasdiLog.debugLog("ProjectResource.changeActiveProject( " + "changing the active project of the user to " + sProjectId + " failed");
-			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The changing of the active project failed.")).build();
-		}
+		catch (Exception oEx) {
+			WasdiLog.errorLog("ProjectResource.changeActiveProject: error ", oEx);
+			return Response.serverError().build();
+		}		
 	}
 
 	@DELETE
@@ -390,31 +446,42 @@ public class ProjectResource {
 		User oUser = Wasdi.getUserFromSession(sSessionId);
 
 		if (oUser == null) {
-			WasdiLog.debugLog("ProjectResource.deleteProject: invalid session");
-			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(MSG_ERROR_INVALID_SESSION)).build();
+			WasdiLog.warnLog("ProjectResource.deleteProject: invalid session");
+			return Response.status(Status.UNAUTHORIZED).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_INVALID_SESSION.name())).build();
 		}
 
 		if (Utils.isNullOrEmpty(sProjectId)) {
-			WasdiLog.debugLog("ProjectResource.deleteProject: invalid project id");
+			WasdiLog.warnLog("ProjectResource.deleteProject: invalid project id");
 			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("Invalid projectId.")).build();
 		}
+		
+		try {
+			ProjectRepository oProjectRepository = new ProjectRepository();
 
-		ProjectRepository oProjectRepository = new ProjectRepository();
+			Project oProject = oProjectRepository.getProjectById(sProjectId);
 
-		Project oProject = oProjectRepository.getProjectById(sProjectId);
+			if (oProject == null) {
+				WasdiLog.warnLog("ProjectResource.deleteProject: project does not exist");
+				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the name exists.")).build();
+			}
+			
+			if (!PermissionsUtils.canUserWriteSubscription(oUser.getUserId(), sProjectId)) {
+				WasdiLog.warnLog("ProjectResource.deleteProject: user cannot write subscription");
+				return Response.status(Status.FORBIDDEN).entity(new ErrorResponse(ClientMessageCodes.MSG_ERROR_CANNOT_ACCESS.name())).build();				
+			}
 
-		if (oProject == null) {
-			WasdiLog.debugLog("ProjectResource.deleteProject: project does not exist");
-			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("No project with the name exists.")).build();
+			if (oProjectRepository.deleteProject(sProjectId)) {
+				return Response.ok(new SuccessResponse(sProjectId)).build();
+			} 
+			else {
+				WasdiLog.debugLog("ProjectResource.deleteProject: deletion failed");
+				return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The deletion of the project failed.")).build();
+			}			
 		}
-
-		if (oProjectRepository.deleteProject(sProjectId)) {
-			return Response.ok(new SuccessResponse(sProjectId)).build();
-		} 
-		else {
-			WasdiLog.debugLog("ProjectResource.deleteProject: deletion failed");
-			return Response.status(Status.BAD_REQUEST).entity(new ErrorResponse("The deletion of the project failed.")).build();
-		}
+		catch (Exception oEx) {
+			WasdiLog.errorLog("ProjectResource.deleteProject: error ", oEx);
+			return Response.serverError().build();
+		}	
 	}
 
 	private static Project convert(ProjectEditorViewModel oProjectEVM) {
